Added note function to Calendar
Calendar now has a note function, but it doesn't save the notes when the app is fully terminated. Also doesn't allow notes for particular dates. I found we need to use an SQLite DB to store the data. It does this using a subclass called ROOM.

Also changed the distance between the top calculator buttons.